primal

+++ database/migrations/2023_08_21_092755_add_digiflazz_saldo_to_app_settings_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('app_settings', function (Blueprint $table) {
-            $table->integer('digiflazz_saldo')->nullable()->after('digiflazz_key');
+            $table->bigInteger('digiflazz_saldo')->nullable()->after('digiflazz_key');
         });
     }
